pyproject.toml
# SPDX-FileCopyrightText: 2024 Carlo Castoldi <carlo.castoldi@outlook.com>
#
# SPDX-License-Identifier: CC0-1.0

[tool.poetry]
name = "braian"
version = "0.1.0"
description = "A python library for whole-brain analysis"
authors = ["Carlo Castoldi <carlo.castoldi@outlook.com>"]
license = "LGPL-3.0-or-later"
readme = "README.md"
packages = [{include = "braian"}]

[tool.poetry.dependencies]
python = ">=3.11,<4"
python-igraph = "^0.10.4"
numpy = "^1.24.3"
pandas = "^2.0.1"
plotly = "^5.14.1"
scipy = "^1.10.1"
kaleido = "0.2.1"               # used to export plotly images. "0.2.1.post1" version can't be installed
beautifulsoup4 = "^4.12.2"      # used to blacklist regions from Allen's ontology that have no annotation
brainglobe-heatmap = "^v0.5.2"  # used to visualize brain heatmap for IEGs overlaps
urllib3 = "^1.26.16"            # fix poetry+conda bug: https://github.com/python-poetry/poetry/issues/7936
tables = "^3.9.2"               # fix problem with h5py installation
pywin32 = {version = "^306", platform = "win32"}
toml = "^0.10.2"                # used by BraiAnConfig to load TOML file into OrderedDict
pyyaml = "^6.0.2"

[tool.poetry.group.docs]
optional = true

[tool.poetry.group.docs.dependencies]
black = "^24.4.2"
mkdocs = {git = "https://github.com/carlocastoldi/mkdocs.git", rev = "on-template"}
mkdocstrings = {extras = ["python"], version = "^0.25.1"}
mkdocs-plugin-progress = "^1.2.0"
mkdocs-material = "^9.5.25"
mkdocs-jupyter = "^0.24.7"
mkdocs-include-markdown-plugin = "^6.1.1"
mkdocs-multi-theme-plugin = {git = "https://github.com/carlocastoldi/mkdocs-multi-theme-plugin.git", rev = "master"}
markdown-katex = "^202406.1035"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
